Address nits (LinkedInService to /src and minor spacing)
