+++ components/server/ee/src/container-module.ts
@@ -62,7 +62,6 @@ import { EntitlementServiceLicense } from "./billing/entitlement-service-license
 import { EntitlementServiceImpl } from "./billing/entitlement-service";
 import { EntitlementServiceUBP } from "./billing/entitlement-service-ubp";
 import { UsageService, UsageServiceImpl, NoOpUsageService } from "../../src/user/usage-service";
-import { LinkedInService } from "./user/linkedin-service";
 
 export const productionEEContainerModule = new ContainerModule((bind, unbind, isBound, rebind) => {
     rebind(Server).to(ServerEE).inSingletonScope();
@@ -136,5 +135,4 @@ export const productionEEContainerModule = new ContainerModule((bind, unbind, is
             return new NoOpUsageService();
         })
         .inSingletonScope();
-    bind(LinkedInService).toSelf().inSingletonScope();
 });
